Dont use the LIST_INIT macro no need LIST_INIT does no mem checking. Move forward/rewind to hash to func that is used by version checking When version mismatches rew/ff to last head if its removed find next hash ignoreing any possible new additions.
List loops are no longer required to be locked the loop holds a ref
to the list and a ref to returned data the ref to data must be released
in the code.

if the list structure changes even if the "head" is removed the list is 
iterated to find the next position if there is no need [versions match]
this step is not run.

using the loop remove current func is safe as the version is updated
on removal and this is a safe operation.



git-svn-id: http://ns2/svnsrc/taploop/trunk@1264 a03fbada-158d-47ef-9f12-77e997805b15